+++ lib/src/controller.dart
@@ -1,0 +1,9 @@
+// Copyright 2021 Grey & Ivy Inc. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+///
+export 'package:mvc_application/controller.dart' hide AppController, runApp;
+
+/// The Controller for the App itself
+export 'package:playhouse/src/app/controller/app_controller.dart';