+++ lib/src/model.dart
@@ -1,0 +1,17 @@
+// Copyright 2021 Grey & Ivy Inc. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+/// MVC framework
+export 'package:mvc_application/model.dart';
+
+/// Firebase Firestore
+export 'package:dbutils/firestore_db.dart';
+
+/// Assets for testing
+export 'package:playhouse/src/home/model/assets.dart';
+
+/// Firebase database
+export 'package:playhouse/src/home/model/database.dart';
+
+